+++ entrypoints/popup/router.tsx
@@ -370,9 +370,9 @@ export const router = createHashRouter([
                   {
                     path: "zkas/settings",
                     element: (
-                      <ZKasOnly>
+                      <ZKasAvailable>
                         <ZKasSettings />
-                      </ZKasOnly>
+                      </ZKasAvailable>
                     ),
                   },
                   {